requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai>=1.0.0
requests==2.32.3
gunicorn==22.0.0
google-cloud-bigquery>=3.20.0
google-cloud-translate>=3.15.0
flask-limiter>=3.8.0
google-cloud-firestore>=2.15.0
